refactor: extract PDF visual source mapping
